Develop

v 2.0.0 릴리즈

+++ backend/routes/user.js
@@ -1,3 +1,5 @@
+import path from "path";
+import fs from "fs";
 import express from "express";
 import bcrypt from "bcrypt";
 
@@ -7,11 +9,10 @@ import db from "../models/index.js";
 const { User, Image, Post } = db;
 
 const router = express.Router();
+const __dirname = path.resolve();
 
 // 로그인한 유저 정보 가져오기
-router.get("/me", async (req, res, next) => {
-  if (!req.user) return res.status(200).json({ message: "로그인하지 않았습니다.", user: null });
-
+router.get("/me", isLoggedIn, async (req, res, next) => {
   try {
     const fullUser = await User.findOne({
       attributes: ["_id", "name", "provider"],
@@ -64,7 +65,7 @@ router.get("/me/detail", isLoggedIn, async (req, res, next) => {
 
 // 회원가입
 router.post("/", isNotLoggedIn, async (req, res, next) => {
-  const { id, password, name, phone, birthday, imageName } = req.body;
+  const { id, password, name, email, phone, birthday, about, imageName } = req.body;
 
   try {
     const exUser = await User.findOne({ where: { id } });
@@ -77,14 +78,21 @@ router.post("/", isNotLoggedIn, async (req, res, next) => {
       id,
       password: hashedPassword,
       name,
+      email,
       phone,
       birthday,
+      about,
     });
 
     await Image.create({
       name: imageName,
       UserId: createdUser._id,
     });
+
+    // 2022/01/23 - 게시글 생성 완료 시 추가한 이미지 위치 이동 - by 1-blue
+    const oldPath = path.join(__dirname, "public", "images", "preview", imageName);
+    const newPath = path.join(__dirname, "public", "images", imageName);
+    fs.rename(oldPath, newPath, () => {});
 
     return res.status(200).json({ message: `${name}님 회원가입이 완료되었습니다.\n로그인페이지로 이동합니다.` });
   } catch (error) {
@@ -94,13 +102,13 @@ router.post("/", isNotLoggedIn, async (req, res, next) => {
 });
 
 // 2021/12/31 - 특정 유저 정보 가져오기 - by 1-blue
-router.get("/:UserId", isLoggedIn, async (req, res, next) => {
+router.get("/:UserId", async (req, res, next) => {
   const UserId = +req.params.UserId;
 
   try {
     const targetUser = await User.findOne({
       where: { _id: UserId },
-      attributes: ["_id", "name"],
+      attributes: ["_id", "name", "about"],
       include: [
         {
           model: Image,
@@ -131,23 +139,27 @@ router.get("/:UserId", isLoggedIn, async (req, res, next) => {
 
     if (!targetUser) return res.status(404).json({ message: "유저가 존재하지 않습니다." });
 
-    return res.status(200).json({ message: "특정 유저의 정보를 가져오는데 성공했습니다.", user: targetUser });
+    return res
+      .status(200)
+      .json({ message: `${targetUser.name}님의 정보를 가져오는데 성공했습니다.`, user: targetUser });
   } catch (error) {
     console.error("GET /user/:UserId error >> ", error);
     return next(error);
   }
 });
 
-// 2022/01/03 - 로그인한 유저의 기본 정보 변경 - by 1-blue
+// 2022/01/24 - 로그인한 유저의 기본 정보 변경 - by 1-blue
 router.put("/", isLoggedIn, async (req, res, next) => {
-  const { name, phone, birthday, profileImage } = req.body;
+  const { name, email, phone, birthday, about, profileImage } = req.body;
 
   try {
     await User.update(
       {
         name,
+        email,
         phone,
         birthday,
+        about,
       },
       {
         where: {
@@ -163,11 +175,21 @@ router.put("/", isLoggedIn, async (req, res, next) => {
         },
         { where: { UserId: req.user._id } },
       );
+
+      // 2022/01/23 - 게시글 생성 완료 시 추가한 이미지 위치 이동 - by 1-blue
+      const oldPath = path.join(__dirname, "public", "images", "preview", profileImage);
+      const newPath = path.join(__dirname, "public", "images", profileImage);
+      fs.rename(oldPath, newPath, () => {});
     }
 
     res.status(200).json({
-      message: "유저의 정보변경에 성공했습니다.",
-      result: { name, phone, birthday, profileImage },
+      message: `${name}님의 정보를 성공적으로 변경했습니다.`,
+      name,
+      email,
+      phone,
+      birthday,
+      about,
+      profileImage,
     });
   } catch (error) {
     console.error("PUT /user error >> ", error);
@@ -175,16 +197,16 @@ router.put("/", isLoggedIn, async (req, res, next) => {
   }
 });
 
-// 2022/01/03 - 로그인한 유저의 비밀번호 변경 - by 1-blue
+// 2022/01/24 - 로그인한 유저의 비밀번호 변경 - by 1-blue
 router.patch("/", isLoggedIn, async (req, res, next) => {
-  const { prevPassword, currPassword } = req.body;
+  const { currentPassword, password } = req.body;
 
   try {
-    if (!(await bcrypt.compare(prevPassword, req.user.password))) {
-      return res.status(202).json({ message: "비밀번호가 불일치합니다." });
+    if (!(await bcrypt.compare(currentPassword, req.user.password))) {
+      return res.status(202).json({ message: "기존 비밀번호와 불일치합니다." });
     }
 
-    const hashedPassword = await bcrypt.hash(currPassword, 6);
+    const hashedPassword = await bcrypt.hash(password, 6);
 
     await User.update(
       {
@@ -199,7 +221,7 @@ router.patch("/", isLoggedIn, async (req, res, next) => {
     req.session.destroy();
     res
       .status(200)
-      .clearCookie("auth-bluegram")
+      .clearCookie("auth-blegram")
       .json({ message: "비밀번호 변경에 성공하셨습니다.\n강제로 로그아웃되며 로그인페이지로 이동합니다." });
   } catch (error) {
     console.error("PATCH /user error >> ", error);
@@ -207,19 +229,26 @@ router.patch("/", isLoggedIn, async (req, res, next) => {
   }
 });
 
-// 2022/01/03 - 로그인한 유저 회원탈퇴 - by 1-blue
-router.delete("/", isLoggedIn, async (req, res, next) => {
+// 2022/01/24 - 로그인한 유저 회원탈퇴 - by 1-blue
+// 어차피 회원 탈퇴니까 password를 params형태로 넘겨도 상관없다고 판단함
+router.delete("/:password", isLoggedIn, async (req, res, next) => {
+  const { password } = req.params;
+
   try {
+    if (!(await bcrypt.compare(password, req.user.password))) {
+      return res.status(202).json({ message: "기존 비밀번호와 불일치합니다." });
+    }
+
     await User.destroy({ where: { _id: req.user._id } });
 
     req.logout();
     req.session.destroy();
     res
       .status(200)
-      .clearCookie("auth-bluegram")
+      .clearCookie("auth-blegram")
       .json({ message: "회원탈퇴에 성공하셨습니다.\n강제로 로그아웃되며 회원가입페이지로 이동합니다." });
   } catch (error) {
-    console.error("DELETE /user error >> ", error);
+    console.error("DELETE /user/:password error >> ", error);
     return next(error);
   }
 });
